rename reducer file to createReducer export as default to maintain compatibility.
