+++ src/index.js
@@ -9,7 +9,10 @@ module.exports = {
     return require('./components/NetworkConsumer').default;
   },
   get reducer() {
-    return require('./redux/reducer').createReducer;
+    return require('./redux/createReducer').default();
+  },
+  get createReducer() {
+    return require('./redux/createReducer').default;
   },
   get createNetworkMiddleware() {
     return require('./redux/createNetworkMiddleware').default;
